fix #2: playground support colorized output
it uses a third party library (ace) to implement
the tokenization, colorization, ui behaviors.
The original text editor was muted but its javascript code
was kept for compatibility.
The new renderer copies the source code to compile into
a new hidden textarea that is used to communicate with the
original javascript code.