fix #3: cross browser rendering tests were added
WIP.
Under the hood it uses the golang chromedp package.
The tests aim to provide support for major browsers.
It currently works with chrome.
Testing on edge is unfortunately not yet available
as this software won t run on the systems provided by
appveyor.
Safari and opera were not yet implemented.
Firefox should work appropriately.
The test realize screenshots of the ui under
various states, it then commit/add those pictures
in the gh-pages branch.